fix: change workflow start node output when change file config
