+++ projects/app/src/pages/app/detail/components/WorkflowComponents/Flow/nodes/NodeSystemConfig.tsx
@@ -19,6 +19,8 @@ import VariableEdit from '@/components/core/app/VariableEdit';
 import { AppContext } from '@/pages/app/detail/components/context';
 import WelcomeTextConfig from '@/components/core/app/WelcomeTextConfig';
 import FileSelect from '@/components/core/app/FileSelect';
+import { FlowNodeTypeEnum } from '@fastgpt/global/core/workflow/node/constant';
+import { userFilesInput } from '@fastgpt/global/core/workflow/template/system/workflowStart';
 
 type ComponentProps = {
   chatConfig: AppChatConfigType;
@@ -218,6 +220,10 @@ function QuestionInputGuide({ chatConfig: { chatInputGuide }, setAppDetail }: Co
 }
 
 function FileSelectConfig({ chatConfig: { fileSelectConfig }, setAppDetail }: ComponentProps) {
+  const onChangeNode = useContextSelector(WorkflowContext, (v) => v.onChangeNode);
+  const nodes = useContextSelector(WorkflowContext, (v) => v.nodes);
+  const workflowStartNode = nodes.find((item) => item.type === FlowNodeTypeEnum.workflowStart)!;
+
   return (
     <FileSelect
       value={fileSelectConfig}
@@ -229,6 +235,27 @@ function FileSelectConfig({ chatConfig: { fileSelectConfig }, setAppDetail }: Co
             fileSelectConfig: e
           }
         }));
+
+        // Dynamic add or delete userFilesInput
+        const canUploadFiles = e.canSelectFile || e.canSelectImg;
+        const repeatKey = workflowStartNode?.data.outputs.find(
+          (item) => item.key === userFilesInput.key
+        );
+        if (canUploadFiles) {
+          !repeatKey &&
+            onChangeNode({
+              nodeId: workflowStartNode.id,
+              type: 'addOutput',
+              value: userFilesInput
+            });
+        } else {
+          repeatKey &&
+            onChangeNode({
+              nodeId: workflowStartNode.id,
+              type: 'delOutput',
+              key: userFilesInput.key
+            });
+        }
       }}
     />
   );
