+++ projects/app/src/pages/app/detail/components/WorkflowComponents/Flow/nodes/NodeWorkflowStart.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useMemo } from 'react';
+import React from 'react';
 import { NodeProps } from 'reactflow';
 import NodeCard from './render/NodeCard';
 import { FlowNodeItemType } from '@fastgpt/global/core/workflow/type/node.d';
@@ -17,13 +17,11 @@ import {
 } from '@fastgpt/global/core/workflow/node/constant';
 import { WorkflowIOValueTypeEnum } from '@fastgpt/global/core/workflow/constants';
 import { AppContext } from '@/pages/app/detail/components/context';
-import { userFilesInput } from '@fastgpt/global/core/workflow/template/system/workflowStart';
 
 const NodeStart = ({ data, selected }: NodeProps<FlowNodeItemType>) => {
   const { t } = useTranslation();
   const { nodeId, outputs } = data;
   const nodeList = useContextSelector(WorkflowContext, (v) => v.nodeList);
-  const onChangeNode = useContextSelector(WorkflowContext, (v) => v.onChangeNode);
   const { appDetail } = useContextSelector(AppContext, (v) => v);
 
   const variablesOutputs = useCreation(() => {
@@ -55,30 +53,6 @@ const NodeStart = ({ data, selected }: NodeProps<FlowNodeItemType>) => {
       };
     });
   }, [nodeList, t]);
-
-  // Dynamic add or delete userFilesInput
-  useEffect(() => {
-    const canUploadFiles =
-      appDetail.chatConfig?.fileSelectConfig?.canSelectFile ||
-      appDetail.chatConfig?.fileSelectConfig?.canSelectImg;
-    const repeatKey = outputs.find((item) => item.key === userFilesInput.key);
-
-    if (canUploadFiles) {
-      !repeatKey &&
-        onChangeNode({
-          nodeId,
-          type: 'addOutput',
-          value: userFilesInput
-        });
-    } else {
-      repeatKey &&
-        onChangeNode({
-          nodeId,
-          type: 'delOutput',
-          key: userFilesInput.key
-        });
-    }
-  }, [appDetail.chatConfig?.fileSelectConfig, nodeId, onChangeNode, outputs]);
 
   return (
     <NodeCard
